+++ app/upgrades.go
@@ -18,6 +18,7 @@ import (
 	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
 	upgradetypes "github.com/cosmos/cosmos-sdk/x/upgrade/types"
 
+	appparams "github.com/umee-network/umee/v6/app/params"
 	"github.com/umee-network/umee/v6/util"
 	leveragetypes "github.com/umee-network/umee/v6/x/leverage/types"
 )
@@ -47,6 +48,7 @@ func (app UmeeApp) RegisterUpgradeHandlers() {
 	app.registerOutdatedPlaceholderUpgrade("v6.1")
 	app.registerOutdatedPlaceholderUpgrade("v6.2")
 	app.registerUpgrade("v6.3", upgradeInfo)
+
 	app.registerUpgrade6_4(upgradeInfo)
 }
 
@@ -54,8 +56,12 @@ func (app *UmeeApp) registerUpgrade6_4(_ upgradetypes.Plan) {
 	planName := "v6.4"
 
 	app.UpgradeKeeper.SetUpgradeHandler(planName,
-		func(ctx sdk.Context, plan upgradetypes.Plan, fromVM module.VersionMap) (module.VersionMap, error) {
+		func(ctx sdk.Context, _ upgradetypes.Plan, fromVM module.VersionMap) (module.VersionMap, error) {
 			printPlanName(planName, ctx.Logger())
+			// Add UX denom aliases to metadata
+			app.BankKeeper.SetDenomMetaData(ctx, appparams.UmeeTokenMetadata())
+
+			// migrate leverage token settings
 			tokens := app.LeverageKeeper.GetAllRegisteredTokens(ctx)
 			for _, token := range tokens {
 				// this will allow existing interest rate curves to pass new Token validation
@@ -79,7 +85,7 @@ func (app *UmeeApp) registerUpgrade6_0(upgradeInfo upgradetypes.Plan) {
 	util.Panic(err)
 
 	app.UpgradeKeeper.SetUpgradeHandler(planName,
-		func(ctx sdk.Context, plan upgradetypes.Plan, fromVM module.VersionMap) (module.VersionMap, error) {
+		func(ctx sdk.Context, _ upgradetypes.Plan, fromVM module.VersionMap) (module.VersionMap, error) {
 			printPlanName(planName, ctx.Logger())
 			if err := app.LeverageKeeper.SetParams(ctx, leveragetypes.DefaultParams()); err != nil {
 				return fromVM, err
@@ -100,7 +106,7 @@ func (app *UmeeApp) registerUpgrade4_3(upgradeInfo upgradetypes.Plan) {
 	const planName = "v4.3"
 	app.UpgradeKeeper.SetUpgradeHandler(planName, onlyModuleMigrations(app, planName))
 	app.UpgradeKeeper.SetUpgradeHandler(planName,
-		func(ctx sdk.Context, plan upgradetypes.Plan, fromVM module.VersionMap) (module.VersionMap, error) {
+		func(ctx sdk.Context, _ upgradetypes.Plan, fromVM module.VersionMap) (module.VersionMap, error) {
 			ctx.Logger().Info("Upgrade handler execution", "name", planName)
 
 			// set the ICS27 consensus version so InitGenesis is not run
@@ -145,7 +151,7 @@ func (app *UmeeApp) registerUpgrade4_3(upgradeInfo upgradetypes.Plan) {
 }
 
 func onlyModuleMigrations(app *UmeeApp, planName string) upgradetypes.UpgradeHandler {
-	return func(ctx sdk.Context, plan upgradetypes.Plan, fromVM module.VersionMap) (module.VersionMap, error) {
+	return func(ctx sdk.Context, _ upgradetypes.Plan, fromVM module.VersionMap) (module.VersionMap, error) {
 		printPlanName(planName, ctx.Logger())
 		ctx.Logger().Info("-----------------------------\n-----------------------------")
 		ctx.Logger().Info("Upgrade handler execution", "name", planName)