+++ x/leverage/client/tests/tests.go
@@ -91,7 +91,7 @@ func (s *IntegrationTests) TestLeverageScenario() {
 			Response: &types.QueryRegisteredTokensResponse{},
 			ExpectedResponse: &types.QueryRegisteredTokensResponse{
 				Registry: []types.Token{
-					fixtures.Token(appparams.BondDenom, appparams.DisplayDenom, 6),
+					fixtures.Token(appparams.BondDenom, appparams.LegacyDisplayDenom, 6),
 				},
 			},
 			ErrMsg: "",
@@ -113,7 +113,7 @@ func (s *IntegrationTests) TestLeverageScenario() {
 			Response: &types.QueryRegisteredTokensResponse{},
 			ExpectedResponse: &types.QueryRegisteredTokensResponse{
 				Registry: []types.Token{
-					fixtures.Token(appparams.BondDenom, appparams.DisplayDenom, 6),
+					fixtures.Token(appparams.BondDenom, appparams.LegacyDisplayDenom, 6),
 				},
 			},
 			ErrMsg: "",
@@ -126,12 +126,12 @@ func (s *IntegrationTests) TestLeverageScenario() {
 			},
 			Response: &types.QueryMarketSummaryResponse{},
 			ExpectedResponse: &types.QueryMarketSummaryResponse{
-				SymbolDenom:         "UMEE",
+				SymbolDenom:         appparams.LegacyDisplayDenom,
 				Exponent:            6,
 				OraclePrice:         &oracleSymbolPrice,
 				OracleHistoricPrice: &oracleSymbolPrice,
 				UTokenExchangeRate:  sdk.OneDec(),
-				// Borrow rate * (1.52 - ReserveFactor - OracleRewardFactor)
+				// Borrow rate * (1 - ReserveFactor - OracleRewardFactor)
 				// 1.52 * (1 - 0.2 - 0.01) = 1.2008
 				Supply_APY: sdk.MustNewDecFromStr("1.2008"),
 				// This is an edge case technically - when effective supply, meaning