+++ x/oracle/keeper/keeper_test.go
@@ -28,9 +28,9 @@ import (
 )
 
 const (
-	displayDenom string = appparams.DisplayDenom
-	bondDenom    string = appparams.BondDenom
-	initialPower        = int64(10000000000)
+	displayDenom = appparams.LegacyDisplayDenom
+	bondDenom    = appparams.BondDenom
+	initialPower = int64(10000000000)
 )
 
 // Test addresses