+++ x/uibc/quota/intest/suite_test.go
@@ -28,7 +28,6 @@ import (
 )
 
 const (
-	displayDenom string = appparams.DisplayDenom
 	bondDenom    string = appparams.BondDenom
 	initialPower        = int64(10000000000)
 )